refactor: riot service 클래스 추출
