+++ apps/api/src/app.module.ts
@@ -2,12 +2,14 @@ import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { ConfigModule } from '@nestjs/config';
+import { RiotModule } from './riot/riot.module';
 
 @Module({
   imports: [
     ConfigModule.forRoot({
       isGlobal: true,
     }),
+    RiotModule,
   ],
   controllers: [AppController],
   providers: [AppService],
