+++ apps/api/src/riot/riot.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { RiotService } from './riot.service';
+
+@Module({
+  providers: [RiotService],
+  exports: [RiotService],
+})
+export class RiotModule {}
